+++ src/cli/richlist.go
@@ -9,13 +9,12 @@ import (
 	"github.com/skycoin/skycoin/src/api"
 )
 
-func richlistCmd() cobra.Command {
-	name := "richlist"
-	return cobra.Command{
-		Short:                 name,
+func richlistCmd() *cobra.Command {
+	return &cobra.Command{
+		Short:                 "Get skycoin richlist",
 		Long:                  "Returns top N address (default 20) balances (based on unspent outputs). Optionally include distribution addresses (exluded by default).",
-		Use:                   "[top N addresses (20 default)] [include distribution addresses (false default)]",
-		Args:                  cobra.MinimumNArgs(2),
+		Use:                   "richlist [top N addresses (20 default)] [include distribution addresses (false default)]",
+		Args:                  cobra.MaximumNArgs(2),
 		DisableFlagsInUseLine: true,
 		SilenceUsage:          true,
 		RunE:                  getRichlist,
@@ -23,14 +22,16 @@ func richlistCmd() cobra.Command {
 }
 
 func getRichlist(_ *cobra.Command, args []string) error {
-	num := args[0]
-	if num == "" {
-		num = "20" // default to 20 addresses
-	}
+	// default values
+	num := "20"
+	dist := "false"
 
-	dist := args[1]
-	if dist == "" {
-		dist = "false" // default to false
+	switch len(args) {
+	case 1:
+		num = args[0]
+	case 2:
+		num = args[0]
+		dist = args[1]
 	}
 
 	n, err := strconv.Atoi(num)